nri: fixing access to nri socket
The path to nri socket has changed on latest version from:
/var/run/nri.sock -> /var/run/nri/nri.sock
full commit:https://github.com/cri-o/cri-o/pull/6648/commits/1d0ca6fd19c8318833704bd33609abaac20b97ea

This required changing the volume mount path on the host.
Set the volume type as Socket so next time it won't allow
the pod to start if the socket is missing.

Also removing the privileged permissions which are not needed
for accessing the socket.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>