e2e: infra: setup before e2e run
The e2e test has many prerequisitues before
it can actually test the plugin functionality.

This patch intend to provides an automatic (optional) way
for setup the cluster and prepare it for the e2e tests.

It does the following:
 1. Deploy a MachineConfig to enable NRI
 2. Deploy KubeletConfig to configure reserved CPUs
 3. Deploy the plugin componenets (DaemonSet, ServiceAccount,
    SecurityContextConstraints, etc.)

In addition it's possible to revert the cluster to
its initial state once the test run ends.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>