manifests: add option to set name
When the client controls the name,
it can store the name and delete the objects without having the need to hold the objects themselves.

In our case for example we want to be able to
the node-plugin (delete the objects) without deleting the profile.

This can be done easily by specifying the name
of the objects we want to delete.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>